Split controller.go off of the Certificate reconciler
